rename: useCardValidate, useCardArrayValidate hook 이름을 각각 useInputValidate, useInputArrayValidate로 변경
