+++ hooks/src/lib/index.ts
@@ -1,8 +1,8 @@
-export { default as useCardArrayValidate } from './useCardArrayValidate/useCardArrayValidate';
+export { default as useInputArrayValidate } from './useInputArrayValidate/useInputArrayValidate';
 export { default as useCardCompany } from './useCardCompany/useCardCompany';
 export { default as useCardCVC } from './useCardCVC/useCardCVC';
 export { default as useCardExpirationDate } from './useCardExpirationDate/useCardExpirationDate';
 export { default as useCardHolderName } from './useCardHolderName/useCardHolderName';
 export { default as useCardNumbers } from './useCardNumbers/useCardNumbers';
 export { default as useCardPassword } from './useCardPassword/useCardPassword';
-export { default as useCardValidate } from './useCardValidate/useCardValidate';
+export { default as useInputValidate } from './useInputValidate/useInputValidate';
